Windows support for pachctl. First version.
Certain things might not work, but the thing compiles and runs usefully
under Windows.

Could be advertised as `beta` for Windows :)
