+++ src/server/pfs/cmds/mount_windows.go
@@ -1,0 +1,9 @@
+// +build windows
+
+package cmds
+
+import "github.com/spf13/cobra"
+
+func loadMountCommands(address string, metrics bool) []*cobra.Command {
+	return nil
+}
