+++ src/server/pkg/sync/sync_others.go
@@ -1,0 +1,9 @@
+// +build !windows
+
+package sync
+
+import "syscall"
+
+func mkfifo(path string, mode uint32) error {
+	return syscall.Mkfifo(path, mode)
+}
