+++ src/server/pkg/sync/sync_windows.go
@@ -1,0 +1,7 @@
+// +build windows
+
+package sync
+
+func mkfifo(path string, mode uint32) error {
+	return nil
+}
